faster hessian computation

+++ src/simulation/precision.h
@@ -1,1 +1,1 @@
-using solverPrecision = float;
+using solverPrecision = double;
